+++ app/Http/Controllers/Api/V1/CalendarEventController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\V1;
 
 use App\EventCategory;
 use App\Http\Resources\Calendar;
