+++ app/Http/Controllers/Api/V1/CalendarRendererController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\V1;
 
 use App\Calendar;
 use App\Http\Controllers\Controller;
