+++ app/Http/Controllers/Api/V1/EventCommentController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\V1;
 
 use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
