+++ app/Providers/RouteServiceProvider.php
@@ -77,7 +77,7 @@ class RouteServiceProvider extends ServiceProvider
     {
         Route::prefix('api')
              ->middleware('api')
-             ->group(base_path('routes/api.php'));
+             ->group(base_path('routes/api_v1.php'));
     }
 
     protected function configureRateLimiting()
